feat: clean up code
